Switch to more actionable explanations of the air quality badness.
These explanations from an EPA publication found at https://www.airnow.gov/sites/default/files/2019-02/air-quality-guide_pm_2015.pdf and are the most "actionable" version of the explanations.

Additionally, move the prose descriptions into the HTML and toggle them on and off with CSS. This cuts down the strings in the HTML, reduces the amont of manual DOM twiddling we need to do and reduces the chance that our various tables of the differing levels of badness will get out of sync.