Add metadata and offline storage to pass PWA check
Also uses a smaller geographic query for sensors in order to minimize
startup time, both online and from localstorage.

Small nits fixed:

	- adds a <meta name="theme-color"> for full-bleed color
	- sets manifest `start_url` w/ PWA query params
	- adds name/short_name values for homescreen shortcut
	- caches all non-PA assets, including analytics
	- adds MIT license to package.json

To update the site in future, ensure that new assets are added to the
list in `sw.js` and bump the "version" number in that file.

TODO: more aggressive cache busting for the SW impl.